delete function add

+++ app/Http/Controllers/LeavingController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\User;
 use App\Models\Schedule;
+use App\Models\Room;
+
 
 class LeavingController extends Controller
 {
@@ -20,8 +22,27 @@ class LeavingController extends Controller
     public function leaving(Request $request){
         $session_user_id=$request->session()->get('user_id');
         $session_room_id=$request->session()->get('room_id');
-        User::where('user_id',$session_user_id)->delete();
-        Schedule::where('user_id',$session_user_id)->delete();
-        return redirect(url('/roomlogin'));
+        
+        $user_info=["user_id"=>$session_user_id,
+                    "password"=>$request->password];
+
+        $user=new User;
+        $delete_ok=$user->user_id_pass_serch($user_info);
+        if($delete_ok){
+            User::where('user_id',$session_user_id)->delete();
+            Schedule::where('user_id',$session_user_id)->delete();
+
+            //そのルームidを持つユーザーがいなければルームを除去する
+            $room=User::where('room_id',$session_room_id)->first();
+            //dd($room);
+            if(!isset($room)){
+                Room::where('room_id',$session_room_id)->delete();
+            }
+
+            return redirect(url('/roomlogin'));
+        }
+        
+        return redirect(url('/leaving'));
+        
     }
 }
